Changed the old Motif "tissue expression" property based on LUT to a simple ArrayList with Strings.
This also fixed another problem that was previously undetected, namely that Motifs that actually
had tissue expression properties could not actually be deserialized properly again after serialization
(because the STATIC look-up-tables would not be saved together with the objects in the serialization stream).
This meant that it was not possible to save and restore sessions containing motifs with tissue expression
(well it was possible, but you would get an ArrayIndexOutOfBounds exception if you tried to inspect
the motifs that previously had tissue properties).